[Truffle] Attempt to adopt @MessageResolution
