--- truffle/src/main/java/org/jruby/truffle/interop/ForeignReadNode.java
@@ -1,53 +1,0 @@
-/*
- * Copyright (c) 2013, 2016 Oracle and/or its affiliates. All rights reserved. This
- * code is released under a tri EPL/GPL/LGPL license. You can use it,
- * redistribute it and/or modify it under the terms of the:
- *
- * Eclipse Public License version 1.0
- * GNU General Public License version 2
- * GNU Lesser General Public License version 2.1
- */
-package org.jruby.truffle.interop;
-
-import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.dsl.Cached;
-import com.oracle.truffle.api.dsl.NodeChild;
-import com.oracle.truffle.api.dsl.NodeChildren;
-import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.frame.VirtualFrame;
-import com.oracle.truffle.api.interop.AcceptMessage;
-import com.oracle.truffle.api.nodes.Node;
-import com.oracle.truffle.api.object.DynamicObject;
-import org.jruby.truffle.RubyContext;
-import org.jruby.truffle.RubyLanguage;
-import org.jruby.truffle.language.RubyNode;
-import org.jruby.truffle.language.RubyObjectType;
-import org.jruby.truffle.language.dispatch.CallDispatchHeadNode;
-import org.jruby.truffle.language.dispatch.DispatchHeadNodeFactory;
-import org.jruby.truffle.language.dispatch.DoesRespondDispatchHeadNode;
-import org.jruby.truffle.language.objects.ReadObjectFieldNode;
-import org.jruby.truffle.language.objects.ReadObjectFieldNodeGen;
-
-@AcceptMessage(value = "READ", receiverType = RubyObjectType.class, language = RubyLanguage.class)
-public final class ForeignReadNode extends ForeignReadBaseNode {
-
-    @Child private Node findContextNode;
-    @Child private ForeignReadStringCachingHelperNode helperNode;
-
-    @Override
-    public Object access(VirtualFrame frame, DynamicObject object, Object name) {
-        return getHelperNode().executeStringCachingHelper(frame, object, name);
-    }
-
-    private ForeignReadStringCachingHelperNode getHelperNode() {
-        if (helperNode == null) {
-            CompilerDirectives.transferToInterpreter();
-            findContextNode = insert(RubyLanguage.INSTANCE.unprotectedCreateFindContextNode());
-            final RubyContext context = RubyLanguage.INSTANCE.unprotectedFindContext(findContextNode);
-            helperNode = insert(ForeignReadStringCachingHelperNodeGen.create(context, null, null));
-        }
-
-        return helperNode;
-    }
-
-}
